GA tweaks, SRA fix
-SimpleReflexAgent code wasn't preserved properly when integrating Joy's ModelBasedReflexAgent code so I readded the SRA
-Tweaked the GA and training simulations to run for longer and have more SRAs to hopefully force the RNNs to deal with aggression
-Tried simplifying inputs to just 3 nodes per grid cell with food_area, virus_area, player_area but for some reason the population isn't improving

+++ game/model.py
@@ -103,7 +103,7 @@ class Model:
         """
         return list(self._overlap_chunks(cell))
 
-    def get_chunks(self, bounds: tuple[tuple[int, int], tuple[int, int]]):
+    def get_chunks(self, bounds: tuple[tuple[float, float], tuple[float, float]]):
         """
         Yield all chunks that overlap the given bounds.
         :param bounds: (top_left, bottom_right)
